int baneado

+++ plantillas/DivideEtImpera.cpp
@@ -6,20 +6,20 @@
 #define ss second
 #define pb push_back
 #define mp make_pair
-#define sz(x) int((x).size())
+#define sz(x) ll((x).size())
 #define all(v) (v).begin(),(v).end()
 #define trace(x) cerr << #x << " = " << x << endl
 #define fastio ios_base::sync_with_stdio(0);cin.tie(0)
-#define For(i,a,b) for(int i = int(a); i < int(b); ++i)
-#define roF(i,a,b) for(int i = int(a); i >= int(b); i--)
+#define For(i,a,b) for(ll i = ll(a); i < ll(b); ++i)
+#define roF(i,a,b) for(ll i = ll(a); i >= ll(b); i--)
 using namespace std;
 
 typedef long long ll;
-typedef pair<int,int> ii;
-typedef vector<int> vi;
+typedef pair<ll,ll> ii;
+typedef vector<ll> vi;
 typedef vector<ii> vii;
 
-const int N=1e6;
+const ll N=1e6;
 
 
 int main(){
